Add WillFill Back to the game!
fix: check for duplicate characters if there is a duplicate character allow them to change (Could happen in solo que or the old trick ready up all at once)
fix: allow WillFill to be selected and ready up on match start they will be able to select a new character

If they do not pick a character they will be forced to a random character
This happens in willfill and or duplicated character
They could choose not to pick one and play russian roulette with this.

This Requires fix/Reconection patch to be merged.